Add / precision for folder functions

+++ src/Hug/Sftp/Sftp.php
@@ -3,7 +3,11 @@
 namespace Hug\Sftp;
 
 use phpseclib\Net\SFTP as SecFtp;
+use Hug\HString\HString as HString;
 
+/**
+ *
+ */
 class Sftp
 {
 
@@ -20,7 +24,8 @@ class Sftp
      */
     private static function login($server, $user, $password, $port = 22)
     {
-        $sftp = null;
+        $sftp = false;
+        
         try
         {
         	$sftp = new SecFtp($server, $port);
@@ -33,6 +38,7 @@ class Sftp
         {
             error_log("SFtp::login : " . $e->getMessage());
         }
+
         return $sftp;
     }
 
@@ -49,17 +55,12 @@ class Sftp
     public static function test($server, $user, $password, $port = 22)
     {
         $test = false;
-        try
-        {
-        	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
-			{
-                $test = true;
-            }
+
+    	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
+		{
+            $test = true;
         }
-        catch(Exception $e)
-        {
-            error_log("Sftp::test : " . $e->getMessage());
-        }
+
         return $test;
     }
 
@@ -96,7 +97,7 @@ class Sftp
 	}
 
     /**
-     * Deletes a file on remote FTP server
+     * Delete a file on remote SFTP server
      *
      * @param string $server 
      * @param string $user
@@ -126,7 +127,7 @@ class Sftp
         }
         catch(Exception $e)
         {
-            error_log("Sftp delete : " . $e->getMessage());
+            error_log("Sft::delete : " . $e->getMessage());
         }
 
         return $deleted;
@@ -135,6 +136,9 @@ class Sftp
     /**
      * Recursively deletes files and folder in given directory
      *
+	 * If remote_path ends with a slash delete folder content
+	 * otherwise delete folder itself
+	 *
      * @param string $server 
      * @param string $user
      * @param string $password
@@ -147,24 +151,35 @@ class Sftp
     public static function rmdir($server, $user, $password, $remote_path, $port = 22)
     {
         $deleted = false;
+
         try
         {
             if(false !== $sftp = Sftp::login($server, $user, $password, $port))
 			{
-                # Delete
+                # Delete directory content
                 if(Sftp::clean_dir($remote_path, $sftp))
                 {
-                    if($sftp->rmdir($remote_path))
-                    {
-                        $deleted = true;
-                    }
+                	# If remote_path do not ends with /
+			    	if(!HString::ends_with($remote_path, '/'))
+			    	{
+	                	# Delete directory itself
+	                    if($sftp->rmdir($remote_path))
+	                    {
+	                        $deleted = true;
+	                    }
+			    	}
+			    	else
+			    	{
+			    		$deleted = true;
+			    	}
                 }
             }
         }
         catch(Exception $e)
         {
-            error_log("Ftp rmdir : " . $e->getMessage());
+            error_log("Sftp::rmdir : " . $e->getMessage());
         }
+
         return $deleted;
     }
 
@@ -180,15 +195,15 @@ class Sftp
     {
         $clean = false;
 
-        $list = $sftp->nlist($remote_path);
         $to_delete = 0;
         $deleted = 0;
+
+        $list = $sftp->nlist($remote_path);
         foreach ($list as $element)
         {
             if($element!=='.' && $element!=='..')
             {
                 $to_delete++;
-                // error_log('element : ' . $element);
                 
                 if($sftp->is_dir($remote_path . DIRECTORY_SEPARATOR . $element))
                 {
@@ -211,15 +226,20 @@ class Sftp
                 }
             }
         }
+
         if($deleted===$to_delete)
         {
             $clean = true;   
         }
+
         return $clean;
     }
 
 	/**
 	 * Recursively copy files and folders on remote SFTP server
+	 *
+	 * If local_path ends with a slash upload folder content
+	 * otherwise upload folder itself
 	 *
 	 * @param string $server 
 	 * @param string $user
@@ -234,24 +254,33 @@ class Sftp
 	public static function upload_dir($server, $user, $password, $local_path, $remote_path, $port = 22)
 	{
 	    $uploaded = false;
+
 	    try
 	    {
 	    	# Remove trailing slash
-	 		$remote_path = rtrim($remote_path, DIRECTORY_SEPARATOR);   	
-	    	
+	 		$remote_path = rtrim($remote_path, DIRECTORY_SEPARATOR);
+	 		
 	    	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
 			{
+				# If local_path do not ends with /
+		    	if(!HString::ends_with($local_path, '/'))
+		    	{
+			 		# Create fisrt level directory on remote filesystem
+		    		$remote_path = $remote_path . DIRECTORY_SEPARATOR . basename($local_path);
+		    		$sftp->mkdir($remote_path);
+		    	}
+
 				if($sftp->is_dir($remote_path))
 				{
-		            $copy_result = Sftp::upload_all($sftp, $local_path, $remote_path); 
-		            $uploaded = true;
+		            $uploaded = Sftp::upload_all($sftp, $local_path, $remote_path);
 				}
 	        }
 	    }
 	    catch(Exception $e)
 	    {
-	        error_log("Problème SFTP : " . $e->getMessage());
+	        error_log("Sftp::upload_dir : " . $e->getMessage());
 	    }
+
 	    return $uploaded;
 	}
 
@@ -267,55 +296,70 @@ class Sftp
 	 */
 	private static function upload_all($sftp, $local_dir, $remote_dir)
 	{
+		$uploaded_all = false;
 	    try
 	    {
-	    	// error_log('sftp_copy_all ' . $local_dir . ' -> ' . $remote_dir);
+	    	# Create remote directory
 	        if(!$sftp->is_dir($remote_dir))
 	        {
 	            if(!$sftp->mkdir($remote_dir))
 	            {
-	            	error_log("Cannot create directory " . $remote_dir);
-	            	return false;
+	            	throw new Exception("Cannot create remote directory.", 1);
 	            }
-	            // error_log("upload_all : ".$remote_dir." Already exists.");
 	        }
+
+	        $to_upload = 0;
+	        $uploaded = 0;
 
 	        $d = dir($local_dir);
 	        while($file = $d->read())
 	        {
 	            if ($file != "." && $file != "..")
 	            {
+	            	$to_upload++;
+
 	                if(is_dir($local_dir . DIRECTORY_SEPARATOR . $file))
 	                {
-	                	# recursive part
-	                	// error_log('DIR : ' . $local_dir . DIRECTORY_SEPARATOR . $file  .' -> ' . $remote_dir . DIRECTORY_SEPARATOR . $file);
-	                    Sftp::upload_all(
+	                	# Upload directory
+	                	# Recursive part
+	                    if(Sftp::upload_all(
 	                    	$sftp, 
 	                    	$local_dir . DIRECTORY_SEPARATOR . $file, 
-	                    	$remote_dir . DIRECTORY_SEPARATOR . $file);
+	                    	$remote_dir . DIRECTORY_SEPARATOR . $file))
+	                    {
+	                    	$uploaded++;
+	                    }
 	                }
 	                else
-	                { 
-	                	error_log('FILE : ' . $local_dir . DIRECTORY_SEPARATOR . $file  .' -> ' . $remote_dir . DIRECTORY_SEPARATOR . $file);
-	                    $put = $sftp->put(
+	                {
+	                	# Upload file
+	                    if($sftp->put(
 	                    	$remote_dir . DIRECTORY_SEPARATOR . $file,
 	                    	$local_dir . DIRECTORY_SEPARATOR . $file,
-	                    	SecFtp::SOURCE_LOCAL_FILE);
-	                    error_log('put : ' . $put);
+	                    	SecFtp::SOURCE_LOCAL_FILE))
+	                    {
+	                    	$uploaded++;
+	                    }
 	                } 
 	            }
 	        }
-	        $d->close(); 
+	        $d->close();
+
+	        if($to_upload===$uploaded)
+	        {
+				$uploaded_all = true;	        	
+	        }
 	    }
 	    catch(Exception $e)
 	    {
-	        error_log("Sftp::upload_all : ".$e);
+	        error_log("Sftp::upload_all : " . $e->getMessage());
 	    }
-	    return true;
+
+	    return $uploaded_all;
 	}
 
 	/**
-     * Downloads a file from remote SFTP server
+     * Download a file from remote SFTP server
      *
      * @param string $server 
      * @param string $user
@@ -330,6 +374,7 @@ class Sftp
     public static function download($server, $user, $password, $remote_file, $local_file, $port = 22)
     {
         $downloaded = false;
+
         try
         {
         	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
@@ -343,13 +388,17 @@ class Sftp
         }
         catch(Exception $e)
         {
-            error_log("Sftp download : " . $e->getMessage());
+            error_log("Sftp::download : " . $e->getMessage());
         }
+
         return $downloaded;
     }
 
     /**
-     * Downloads a directory from remote FTP server
+     * Download a directory from remote SFTP server
+     *
+     * If remote_dir ends with a slash download folder content
+     * otherwise download folder itself
      *
      * @param string $server 
      * @param string $user
@@ -364,18 +413,26 @@ class Sftp
     public static function download_dir($server, $user, $password, $remote_dir, $local_dir, $port = 22)
     {
         $downloaded = false;
+
         try
         {
             if(is_dir($local_dir) && is_writable($local_dir))
             {
                 if(false !== $sftp = Sftp::login($server, $user, $password, $port))
 				{
-					# Create fisrt level directory on local filesystem
-					$local_dir = rtrim($local_dir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . basename($remote_dir);
-					if(mkdir($local_dir))
-					{
-	                    $downloaded = Sftp::download_all($sftp, $remote_dir, $local_dir);
-					}
+					# If local_path do not ends with /
+			    	if(!HString::ends_with($remote_dir, '/'))
+			    	{
+				 		# Create fisrt level directory on local filesystem
+						$local_dir = rtrim($local_dir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . basename($remote_dir);
+			    		mkdir($local_dir);
+			    	}
+					
+			    	# Remove trailing slash
+			 		$local_dir = rtrim($local_dir, DIRECTORY_SEPARATOR);
+
+					# Recursive part
+                    $downloaded = Sftp::download_all($sftp, $remote_dir, $local_dir);
                 }
             }
             else
@@ -385,19 +442,20 @@ class Sftp
         }
         catch(Exception $e)
         {
-            error_log("Ftp download_dir : " . $e->getMessage());
+            error_log("Sftp::download_dir : " . $e->getMessage());
         }
+
         return $downloaded;
     }
 
     /**
-     *
+     * Recursive function to download remote files
      *
      * @param ressource $sftp
      * @param string $remote_dir
      * @param string $local_dir
      *
-     * @return bool true
+     * @return bool $downloaded
      *
      */
     private static function download_all($sftp, $remote_dir, $local_dir)
@@ -458,25 +516,26 @@ class Sftp
         }
         catch(Exception $e)
         {
-            error_log("Ftp download_all : " . $e->getMessage());
+            error_log("Sftp::download_all : " . $e->getMessage());
         }
+
         return $download_all;
     }
 
 	/**
-     * Renames a file on remote SFTP server
+     * Rename a file on remote SFTP server
      *
      * @param string $server 
      * @param string $user
      * @param string $password
-     * @param string $old_file
-     * @param string $new_file
+     * @param string $current_filename
+     * @param string $new_filename
      * @param int $port
      *
      * @return bool $renamed
      *
      */
-	public static function rename($server, $user, $password, $old_file, $new_file, $port = 22)
+	public static function rename($server, $user, $password, $current_filename, $new_filename, $port = 22)
 	{
 	    $renamed = false;
 
@@ -484,7 +543,7 @@ class Sftp
 	    {
 	    	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
 			{
-				if($sftp->rename($old_file, $new_file))
+				if($sftp->rename($current_filename, $new_filename))
 				{
 	                $renamed = true;
 				}
@@ -499,7 +558,7 @@ class Sftp
 	}
 
 	/**
-     * Creates a directory on remote SFTP server
+     * Create a directory on remote SFTP server
      *
      * @param string $server 
      * @param string $user
@@ -533,7 +592,7 @@ class Sftp
 	}
 
 	/**
-     * Creates a file on remote SFTP server
+     * Create and fill in a file on remote SFTP server
      *
      * @param string $server 
      * @param string $user
@@ -544,9 +603,10 @@ class Sftp
      * @return bool $content
      *
      */
-    public static function touch($server, $user, $password, $remote_file, $content, $port = 22)
+    public static function touch($server, $user, $password, $remote_file, $content = '', $port = 22)
     {
         $created = false;
+
         try
         {
             if(false !== $sftp = Sftp::login($server, $user, $password, $port))
@@ -566,11 +626,12 @@ class Sftp
         {
             error_log("Sftp::touch : " . $e->getMessage());
         }
+
         return $created;
     }
 
 	/**
-	 * Uploads a file on SFTP server
+	 * Upload a file on SFTP server
 	 *
 	 * @param string $server 
 	 * @param string $user
@@ -579,24 +640,17 @@ class Sftp
 	 * @param string $remote_file
 	 * @param int $port
 	 *
-	 * @return array $uploaded Result : status (error|success) & message (explains error's origin : most probably access problem)
+	 * @return bool $uploaded
 	 *
 	 */
-	public static function upload($server, $user, $password, $local_file, $remote_file = '', $port = 22)
+	public static function upload($server, $user, $password, $local_file, $remote_file, $port = 22)
 	{
 	    $uploaded = false;
 	    
 	    try
 	    {
-	    	# SET ROOT PATH AND SAME FILE NAME
-		    if($remote_file==='')
-		    {
-		        $remote_file = basename($local_file);
-		    }
-
 	    	if(false !== $sftp = Sftp::login($server, $user, $password, $port))
 			{
-				// error_log('Login SFTP pwd : ' . $sftp->pwd());
 				if($sftp->put($remote_file, $local_file, SecFtp::SOURCE_LOCAL_FILE))
 				{
 	                $uploaded = true;
@@ -605,13 +659,14 @@ class Sftp
 	    }
 	    catch(Exception $e)
 	    {
-	        error_log("Problème FTP : ".$e);
+	        error_log("Sftp::upload : " . $e->getMessage());
 	    }
+
 	    return $uploaded;
 	}
 
 	/**
-     * List files on SFTP server
+     * List files in given directory on SFTP server
      *
      * @param string $server 
      * @param string $user
@@ -619,7 +674,7 @@ class Sftp
      * @param string $path
      * @param int $port
      *
-     * @return array $uploaded Result : status (error|success) & message (explains error's origin : most probably access problem)
+     * @return array $files Files listed in directory or false
      *
      */
     public static function scandir($server, $user, $password, $path, $port = 22)
@@ -647,7 +702,7 @@ class Sftp
      * @param string $password
      * @param int $port
      *
-     * @return array $uploaded Result : status (error|success) & message (explains error's origin : most probably access problem)
+     * @return string $dir Print Working Directory or false
      *
      */
     public static function pwd($server, $user, $password, $port = 22)
